+++ tests/testthat/test-GDALRaster-class.R
@@ -19,11 +19,11 @@ test_that("infoAsJSON() returns string output", {
 
 test_that("dataset parameters are correct", {
 	lcp_file <- system.file("extdata/storm_lake.lcp", package="gdalraster")
-	ds <- new(GDALRaster, lcp_file, TRUE)
+	ds <- new(GDALRaster, lcp_file)
 	expect_length(ds$getFileList(), 2)
 	ds$close()
 	evt_file <- system.file("extdata/storml_evt.tif", package="gdalraster")
-	ds <- new(GDALRaster, evt_file, TRUE)
+	ds <- new(GDALRaster, evt_file)
 	expect_equal(ds$getDriverShortName(), "GTiff")
 	expect_equal(ds$getDriverLongName(), "GeoTIFF")
 	expect_equal(ds$getRasterXSize(), 143)
@@ -34,6 +34,21 @@ test_that("dataset parameters are correct", {
 	expect_equal(ds$res(), c(30,30))
 	expect_equal(ds$dim(), c(143,107,1))
 	ds$close()
+
+	# using dataset open options
+	elev_file <- system.file("extdata/storml_elev.tif", package="gdalraster")
+	oo <- "NUM_THREADS=2"
+	ds_oo <- new(GDALRaster, elev_file, TRUE, oo)
+	expect_equal(ds_oo$getDriverShortName(), "GTiff")
+	expect_equal(ds_oo$getDriverLongName(), "GeoTIFF")
+	expect_equal(ds_oo$getRasterXSize(), 143)
+	expect_equal(ds_oo$getRasterYSize(), 107)
+	expect_equal(ds_oo$getGeoTransform(),
+					c(323476.1,30.0,0.0,5105082.0,0.0,-30.0))
+	expect_equal(ds_oo$bbox(), c(323476.1,5101872.0,327766.1,5105082.0))
+	expect_equal(ds_oo$res(), c(30,30))
+	expect_equal(ds_oo$dim(), c(143,107,1))
+	ds_oo$close()
 })
 
 test_that("band-level parameters are correct", {